Fixed "warning: long unsigned int format, unsigned int arg" in parse_login (login_sql\login.c)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@7941 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/login_sql/login.c
@@ -1767,8 +1767,8 @@ int parse_login(int fd) {
 					}
 				} //End login log of error.
 				if ((result == 1) && (dynamic_pass_failure_ban != 0) && log_login){	// failed password
-					sprintf(tmpsql,"SELECT count(*) FROM `%s` WHERE `ip` = '%lu' AND `rcode` = '1' AND `time` > NOW() - INTERVAL %d MINUTE",
-						loginlog_db,ntohl(ipl), dynamic_pass_failure_ban_time);	//how many times filed account? in one ip.
+					sprintf(tmpsql,"SELECT count(*) FROM `%s` WHERE `ip` = '%u' AND `rcode` = '1' AND `time` > NOW() - INTERVAL %d MINUTE",
+						loginlog_db,(unsigned int)ntohl(ipl), dynamic_pass_failure_ban_time);	//how many times filed account? in one ip.
 					if(mysql_query(&mysql_handle, tmpsql)) {
 						ShowSQL("DB error - %s\n",mysql_error(&mysql_handle));
 						ShowDebug("at %s:%d - %s\n", __FILE__,__LINE__,tmp_sql);
